Unify shared code between IA import & WARC import
